requirements.txt
archspec @ file:///croot/archspec_1697725767277/work
boltons @ file:///work/ci_py311/boltons_1677685195580/work
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///croot/conda_1707839125844/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
cryptography @ file:///croot/cryptography_1702070282333/work
distro @ file:///croot/distro_1701455004953/work
filelock @ file:///croot/filelock_1700591183607/work
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
idna @ file:///work/ci_py311/idna_1676822698822/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
menuinst @ file:///croot/menuinst_1702390294373/work
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///croot/numpy_and_numpy_base_1704311704800/work/dist/numpy-1.26.3-cp311-cp311-linux_x86_64.whl#sha256=10a078151ecec16bafb535f7487635217625fa06536dec8509e514648c78d626
packaging @ file:///croot/packaging_1693575174725/work
pillow @ file:///croot/pillow_1707233021655/work
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1690400202158/work
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
sympy @ file:///croot/sympy_1701397643339/work
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm @ file:///croot/tqdm_1679561862951/work
triton==2.2.0
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
urllib3 @ file:///croot/urllib3_1698257533958/work
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
